backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
pydantic==2.11.0
pydantic-settings==2.9.1
python-dotenv==1.0.1
websockets==15.0.1
passlib[bcrypt]==1.7.4
jwt==1.3.1
redis==5.0.1
